pyproject.toml
[project]
name = "events-manager-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.2",
    "django-celery-beat>=2.9.0",
    "django-constance>=4.3.4",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "redis>=7.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
]

[tool.ruff]
exclude = [".venv", "manage.py"]
line-length = 79

[tool.ruff.lint]
ignore = [
    "D100",
    "D101",
    "D102",
    "D104",
    "D200",
    "D203",
    "D212",
    "D213",
    "RUF001",
    "RUF002",
    "RUF003",
    "D107",
    "TCH",
    "ISC001",
    "COM812",
    "ANN002",
    "ANN003",
    "ANN204",
    "TRY003",
    "EM101",
    "UP046",
    "FBT001",
    "ERA001",
    "D105",
    "TRY400",
    "TRY201",
    "D400",
    "ANN401",
]

select = [
    "ANN",
    "D",
    # Enable isort
    # https://docs.astral.sh/ruff/rules/#isort-i
    "I",
    # Remove unused imports
    # https://docs.astral.sh/ruff/rules/unused-import/
    "F401",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]
